+++ CHANGELOG.md
@@ -4,5 +4,4 @@
 * New required variables `ingress_protocol` and `egress_protocol`
 
 ## v0.1.0 - June 15th 2021
-* New security group module to create a variable number of security groups based on
-  the number of rules.
+* New security group module to create a variable number of security groups based on the number of rules.